services/dispatcher_server/requirements.txt
click~=8.1.3
click-config-file~=0.6.0
grpcio-tools~=1.49.1
protobuf~=4.21.7
toml~=0.10.2
